backend/requirements.txt
Flask>=2.0
Flask-Cors
Flask-Bcrypt
Flask-PyMongo
pymongo
requests
beautifulsoup4
apscheduler
pytesseract
Pillow
numpy
selenium
webdriver-manager
gunicorn
python-dotenv
tzdata
pytz
opencv-python-headless
pypdf
pymupdf
